% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatOp.R
\name{MatOp}
\alias{MatOp}
\title{Calculate the diagonal of \eqn{(A+B)*(C+D)}}
\usage{
MatOp(A, B, C, D)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{C}{A matrix.}

\item{D}{A matrix.}
}
\value{
The diagonal of \eqn{(A+B)\times(C+D)}
}
\description{
Calculate the diagonal of \eqn{(A+B)*(C+D)}
}
