% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatOp.R
\name{MatOp}
\alias{MatOp}
\title{Conduct a matrix operation and print the diagonal elements of the product}
\usage{
MatOp(A, B, C, D)
}
\arguments{
\item{A}{A matrix}

\item{B}{A matrix. Size equals size of A}

\item{C}{A matrix. Number of rows equals A's number of columns}

\item{D}{A matrix. Size equals size of C}
}
\description{
Conduct a matrix operation and print the diagonal elements of the product
}
